requirements_olap.txt
pandas>=1.5.0
numpy>=1.21.0
snowflake-sqlalchemy>=1.4.0
sqlalchemy>=1.4.0
plotly>=5.0.0
streamlit>=1.25.0
openpyxl>=3.0.0
xlrd>=2.0.0 